- Added Opus support on the player's bitrate/codec view - Fixed issue #585 (Songs not resuming if they were paused on app close) - Additionally, fixed progress bar and elapsed/remaining not being preset under the same conditions at #585 - WaveBox folder listings are now stored in temporary tables as opposed to on-disk, seeing as they're already on the disk - Root folder listings are now inserted directly into the main cache - Related, added the addRootFolderToMainCache: method to ISMSRootFoldersLoader - Optimized the subfolders song query for WaveBox - Root folders are now listed in alphabetical order for WaveBox regardless of add date due to a query change

+++ Sub/DatabaseSingleton.m
@@ -66,34 +66,51 @@ LOG_LEVEL_ISUB_DEFAULT
 	{
 		[db executeUpdate:@"PRAGMA cache_size = 1"];
 		
-		if (![db tableExists:@"albumListCache"]) 
-		{
-			[db executeUpdate:@"CREATE TABLE albumListCache (id TEXT PRIMARY KEY, data BLOB)"];
-		}
-		if (![db tableExists:@"albumsCache"]) 
-		{
-			[db executeUpdate:@"CREATE TABLE albumsCache (folderId TEXT, title TEXT, albumId TEXT, coverArtId TEXT, artistName TEXT, artistId TEXT)"];
-			[db executeUpdate:@"CREATE INDEX albumsFolderId ON albumsCache (folderId)"];
-		}
-		if (![db tableExists:@"songsCache"]) 
-		{
-			[db executeUpdate:[NSString stringWithFormat:@"CREATE TABLE songsCache (folderId TEXT, %@)", [ISMSSong standardSongColumnSchema]]];
-			[db executeUpdate:@"CREATE INDEX songsFolderId ON songsCache (folderId)"];
-		}
-        if (![db tableExists:@"albumsCacheCount"])
+        // If this is a WaveBox server, simply let the reset method take care of all this.
+        if ([settingsS.serverType isEqualToString:WAVEBOX])
         {
-            [db executeUpdate:@"CREATE TABLE albumsCacheCount (folderId TEXT, count INTEGER)"];
-            [db executeUpdate:@"CREATE INDEX albumsCacheCountFolderId ON albumsCacheCount (folderId)"];
+            ALog(@"This is a WaveBox server.  Creating temporary tables for folder caches.");
+            [db executeUpdate:@"CREATE TEMPORARY TABLE albumListCache (id TEXT PRIMARY KEY, data BLOB)"];
+            [db executeUpdate:@"CREATE TEMPORARY TABLE albumsCache (folderId TEXT, title TEXT, albumId TEXT, coverArtId TEXT, artistName TEXT, artistId TEXT)"];
+            [db executeUpdate:[NSString stringWithFormat:@"CREATE TEMPORARY TABLE songsCache (folderId TEXT, %@)", [ISMSSong standardSongColumnSchema]]];
+            [db executeUpdate:@"CREATE TEMPORARY TABLE albumsCacheCount (folderId TEXT, count INTEGER)"];
+            [db executeUpdate:@"CREATE TEMPORARY TABLE songsCacheCount (folderId TEXT, count INTEGER)"];
+            [db executeUpdate:@"CREATE TEMPORARY TABLE folderLength (folderId TEXT, length INTEGER)"];
         }
-        if (![db tableExists:@"songsCacheCount"])
+        
+        // Otherwise, we should create these tables if they don't already exist.
+        else
         {
-            [db executeUpdate:@"CREATE TABLE songsCacheCount (folderId TEXT, count INTEGER)"];
-            [db executeUpdate:@"CREATE INDEX songsCacheCountFolderId ON songsCacheCount (folderId)"];
-        }
-        if (![db tableExists:@"folderLength"])
-        {
-            [db executeUpdate:@"CREATE TABLE folderLength (folderId TEXT, length INTEGER)"];
-            [db executeUpdate:@"CREATE INDEX folderLengthFolderId ON folderLength (folderId)"];
+            ALog(@"This is not a WaveBox server.  Checking for folder cache tables and creating as necessary...");
+            if (![db tableExists:@"albumListCache"])
+            {
+                [db executeUpdate:@"CREATE TABLE albumListCache (id TEXT PRIMARY KEY, data BLOB)"];
+            }
+            if (![db tableExists:@"albumsCache"]) 
+            {
+                [db executeUpdate:@"CREATE TABLE albumsCache (folderId TEXT, title TEXT, albumId TEXT, coverArtId TEXT, artistName TEXT, artistId TEXT)"];
+                [db executeUpdate:@"CREATE INDEX albumsFolderId ON albumsCache (folderId)"];
+            }
+            if (![db tableExists:@"songsCache"]) 
+            {
+                [db executeUpdate:[NSString stringWithFormat:@"CREATE TABLE songsCache (folderId TEXT, %@)", [ISMSSong standardSongColumnSchema]]];
+                [db executeUpdate:@"CREATE INDEX songsFolderId ON songsCache (folderId)"];
+            }
+            if (![db tableExists:@"albumsCacheCount"])
+            {
+                [db executeUpdate:@"CREATE TABLE albumsCacheCount (folderId TEXT, count INTEGER)"];
+                [db executeUpdate:@"CREATE INDEX albumsCacheCountFolderId ON albumsCacheCount (folderId)"];
+            }
+            if (![db tableExists:@"songsCacheCount"])
+            {
+                [db executeUpdate:@"CREATE TABLE songsCacheCount (folderId TEXT, count INTEGER)"];
+                [db executeUpdate:@"CREATE INDEX songsCacheCountFolderId ON songsCacheCount (folderId)"];
+            }
+            if (![db tableExists:@"folderLength"])
+            {
+                [db executeUpdate:@"CREATE TABLE folderLength (folderId TEXT, length INTEGER)"];
+                [db executeUpdate:@"CREATE INDEX folderLengthFolderId ON folderLength (folderId)"];
+            }
         }
 	}];
 	
@@ -557,24 +574,55 @@ LOG_LEVEL_ISUB_DEFAULT
 	[self.albumListCacheDbQueue inDatabase:^(FMDatabase *db)
 	{
 		// Drop the tables
-		[db executeUpdate:@"DROP TABLE albumListCache"];
-		[db executeUpdate:@"DROP TABLE albumsCache"];
-		[db executeUpdate:@"DROP TABLE albumsCacheCount"];
-		[db executeUpdate:@"DROP TABLE songsCacheCount"];
-		[db executeUpdate:@"DROP TABLE folderLength"];
+        if ([db tableExists:@"albumListCache"])
+        {
+            [db executeUpdate:@"DROP TABLE albumListCache"];
+        }
+        if ([db tableExists:@"albumsCache"])
+        {
+            [db executeUpdate:@"DROP TABLE albumsCache"];
+        }
+        if ([db tableExists:@"albumsCacheCount"])
+        {
+            [db executeUpdate:@"DROP TABLE albumsCacheCount"];
+        }
+        if ([db tableExists:@"songsCacheCount"])
+        {
+            [db executeUpdate:@"DROP TABLE songsCacheCount"];
+        }
+        if ([db tableExists:@"folderLength"])
+        {
+            [db executeUpdate:@"DROP TABLE folderLength"];
+        }
+        
+        NSString *tableCreateType = @"CREATE TABLE";
+        
+        if ([settingsS.serverType isEqualToString:WAVEBOX])
+        {
+            tableCreateType = @"CREATE TEMPORARY TABLE";
+        }
 		
+        ALog(@"AlbumListCache table type is: %@", tableCreateType);
+        
 		// Create the tables and indexes
-		[db executeUpdate:@"CREATE TABLE albumListCache (id TEXT PRIMARY KEY, data BLOB)"];
-		[db executeUpdate:@"CREATE TABLE albumsCache (folderId TEXT, title TEXT, albumId TEXT, coverArtId TEXT, artistName TEXT, artistId TEXT)"];
-		[db executeUpdate:@"CREATE INDEX albumsFolderId ON albumsCache (folderId)"];
-		[db executeUpdate:[NSString stringWithFormat:@"CREATE TABLE songsCache (folderId TEXT, %@)", [ISMSSong standardSongColumnSchema]]];
-		[db executeUpdate:@"CREATE INDEX songsFolderId ON songsCache (folderId)"];
-		[db executeUpdate:@"CREATE TABLE albumsCacheCount (folderId TEXT, count INTEGER)"];
-		[db executeUpdate:@"CREATE INDEX albumsCacheCountFolderId ON albumsCacheCount (folderId)"];
-		[db executeUpdate:@"CREATE TABLE songsCacheCount (folderId TEXT, count INTEGER)"];
-		[db executeUpdate:@"CREATE INDEX songsCacheCountFolderId ON songsCacheCount (folderId)"];
-		[db executeUpdate:@"CREATE TABLE folderLength (folderId TEXT, length INTEGER)"];
-		[db executeUpdate:@"CREATE INDEX folderLengthFolderId ON folderLength (folderId)"];
+        [db executeUpdate:[NSString stringWithFormat:@"%@ albumListCache (id TEXT PRIMARY KEY, data BLOB)", tableCreateType]];
+        [db executeUpdate:[NSString stringWithFormat:@"%@ albumsCache (folderId TEXT, title TEXT, albumId TEXT, coverArtId TEXT, artistName TEXT, artistId TEXT)", tableCreateType]];
+        [db executeUpdate:[NSString stringWithFormat:@"%@ songsCache (folderId TEXT, %@)", tableCreateType, [ISMSSong standardSongColumnSchema]]];
+        [db executeUpdate:[NSString stringWithFormat:@"%@ albumsCacheCount (folderId TEXT, count INTEGER)", tableCreateType]];
+        [db executeUpdate:[NSString stringWithFormat:@"%@ songsCacheCount (folderId TEXT, count INTEGER)", tableCreateType]];
+        [db executeUpdate:[NSString stringWithFormat:@"%@ folderLength (folderId TEXT, length INTEGER)", tableCreateType]];
+        
+        
+        // Since we're using temporary tables for WaveBox, we really don't care about creating indexes for those tables and this can potentially affect the speed of our inserts.
+        
+        if (![settingsS.serverType isEqualToString:WAVEBOX])
+        {
+            [db executeUpdate:@"CREATE INDEX albumsFolderId ON albumsCache (folderId)"];
+            [db executeUpdate:@"CREATE INDEX songsFolderId ON songsCache (folderId)"];
+            [db executeUpdate:@"CREATE INDEX albumsCacheCountFolderId ON albumsCacheCount (folderId)"];
+            [db executeUpdate:@"CREATE INDEX songsCacheCountFolderId ON songsCacheCount (folderId)"];
+            [db executeUpdate:@"CREATE INDEX folderLengthFolderId ON folderLength (folderId)"];
+        }
 	}];
 }
